+++ src/components/Map.vue
@@ -76,7 +76,7 @@
                         mdi-information
                       </v-icon>
                     </template>
-                    <span>Flood waters exit the stream/river channel and overflow onto a flat surface</span>
+                    <span>Flood waters are overflowing the stream/river channel and into the surrounding area.</span>
                   </v-tooltip>
                 </div>
                 <div class="legendIcon">
@@ -99,7 +99,7 @@
                         mdi-information
                       </v-icon>
                     </template>
-                    <span>Flood waters are flooding a pedestrian greenway/trail/path</span>
+                    <span>Pedestrian greenway/trail/path is underwater.</span>
                   </v-tooltip>
                 </div>
                 <div class="legendIcon">
@@ -122,7 +122,7 @@
                         mdi-information
                       </v-icon>
                     </template>
-                    <span>Low lying areas along roads are flooding</span>
+                    <span>Road is underwater.</span>
                   </v-tooltip>
                 </div>
                 <div class="legendIcon">
@@ -145,7 +145,7 @@
                         mdi-information
                       </v-icon>
                     </template>
-                    <span>Water from a river or stream is under the lowest section of a bridge</span>
+                    <span>Water from the river or stream has reached the bottom of the bridge.</span>
                   </v-tooltip>
                 </div>
                 <div class="legendIcon">
@@ -168,7 +168,7 @@
                         mdi-information
                       </v-icon>
                     </template>
-                    <span>A bridge is flooding</span>
+                    <span>Bridge is underwater.</span>
                   </v-tooltip>
                 </div>
                 <div class="legendIcon">
@@ -191,7 +191,7 @@
                         mdi-information
                       </v-icon>
                     </template>
-                    <span>Structures/facilities are flooding</span>
+                    <span>Water from the river or stream has reached the lowest finished floor of the structure (FEMA Finished Floor Elevation, or FFE).</span>
                   </v-tooltip>
                 </div>
                 <div class="legendIcon">
@@ -213,7 +213,7 @@
                         mdi-information
                       </v-icon>
                     </template>
-                    <span>The FEMA 100-year BFE has been reached</span>
+                    <span>The FEMA 100-year Base Flood Elevation (BFE) has been reached.</span>
                   </v-tooltip>
                 </div>
                 <div class="legendIcon">
@@ -235,7 +235,7 @@
                         mdi-information
                       </v-icon>
                     </template>
-                    <span>Other Reference Points are experiencing flooding (parking lot, campground, railroad, etc).</span>
+                    <span>This Flood Impact Location is experiencing flooding. For more information about this location, open the icon’s pop-up and look for the “Flood Impact Type” description.</span>
                   </v-tooltip>
                 </div>
               </div>
@@ -388,7 +388,7 @@ export default {
       tileProviders: tileProviders,
       streamgageMarkers: [],
       radarLayer: {},
-      radarVisible: false,
+      radarVisible: true,
       aqMarkers: [],
       allRPMarkers: [],
       tideMarkers: [],
@@ -632,6 +632,8 @@ export default {
       self.map.on("zoomend dragend", function () {
         self.currentBounds = self.map.getBounds();
       });
+
+      this.getRadarLayer();
 
       // Emit map object to parent component
       self.getMapObject();