+++ CHANGELOG.md
@@ -19,6 +19,22 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 -
 
+## [v1.0.1](https://github.com/USGS-WiM/Thresholds/releases/tag/v1.0.1) - 10/19/2021
+
+### Added
+
+-
+
+### Changed
+
+- Updated tooltips
+- Updated contact us email
+- Radar layer now visible on site load
+
+### Fixed
+
+-
+
 ## [v1.0.0](https://github.com/USGS-WiM/Thresholds/releases/tag/v1.0.0) - 10/12/2021
 
 ### Added